Feature/1.5.28 upgrade (#15)
* Change default examine chat color to green

For better visibility

* Adjust alignment of icons in fixed 2005 interface style and fix inventory icon

* fishing: add missing spots near farming guild

* [maven-release-plugin] prepare for next development iteration

* prayer: add HD prayer bar

* EmoteClue: Add null checks for stashUnit

Some clues may have item requirements, but no STASH unit, which would
cause NPEs during rendering. This commit also marks stashUnit as
nullable so this issue can hopefully be avoided in the future.

* Add snakeskin gear to skill calculator (#9172)

* worldmap: Fix MEP2 and Shadows of the Storm quest tooltips (#9166)

* api: rename unknownSoundValues1 -> queuedSoundEffectLoops

* Add sounds to devtools

* Add Barbarian Assault quick-start option to menu entry swapper

* api: add method to play sound effects while volume is muted

* metronome: play sounds while volume is muted

* Fix typos in Falo the Bard clues (comma, favourite) (#9185)

* Add support for burning stews to Cooking plugin (#9196) (#9202)

* Fix location of Kourend Agility obstacle on world map (#9203)

Fixes #9176

* config manager: treat null and empty string equally in setDefaultConfiguration

The config manager treats null and empty string as equal and will unset
config values which are set to the empty string with the config client.

This fixes the config manager applying default config values which are the
empty string when the current value is null, causing it to send an
unnecessary unset request

* client: fix behavior of walking where there are no tiles

This only happens with a plugin force setting checkClick. This would
cause the client to possibly incorrectly send a walk packet next tick if
viewportWalking was set, since nothing ever clears it.

* npchighlight: Do not render 'null' names (#9093)

Fix names showing as null in the NPC highlight plugin

Fixes #9084

* Fix WC world map icons around Hosidius area (#9205)

Fixes #9204

* mining plugin: only show timers on same level in motherlode mine

* Added Larran's chest to loot tracker (#9209)

Closes #9177

* interface styles: add RuneScape cross sprites option

* Replace use of deprecated Guava Files write method

* Add tooltips to DevTools buttons (#9224)

* chat commands: add ba high gamble command

* chat commands: add duel arena chat command

* Fix Father Jean's coordinates in easy cryptic clue

* npc indicators: Fix long respawn timer bug

By killing a monster which has a known respawn timer and leaving its
respawn area (by walking away or teleporting) and returning to it some
time after it has respawned, the plugin can erroneously overwrite its
respawn time with a much longer one. This commit fixes that issue by
assuming that the lower of the observed respawn time and the previously
recorded respawn time is the correct one.

* Add ENTER/ESC key events for screenmarker saving

* Clean up some instances of double white space

* Add hover preview to screen markers

* HotColdClue: Refactor to use enums for temperatures

* HotColdClue: Add hot-cold solver class

This adds a general hot-cold puzzle solver class and implements it in
HotColdClue.

* cluescrolls: Add beginner hot-cold clues

* HotColdTemperature: Fix warm-but-colder temperature bug

Because the word "cold" is used both as a temperature ("The device is
cold"), and in a temperature change ("but colder than last time."),
HotColdTemperature tried to interpret COLDER temperature changes as a
COLD temperature match. By matching only the section of the temperature
string preceding the optional comma, this bug is avoided.

* HotColdClue: Improve behavior on visibly shaking location

This change causes the locations list to be narrowed on all locator orb
checks, having the consequence of narrowing the list to the visibly
shaking location (for better display via `makeOverlayHint()'). The call
to `reset()' must be removed as it would re-add all locations to the
instance's solver when finding the dig location, which effectively
reverses the apparent narrowing the overlay displayed until that point.

* HotColdClue: Clear final location on clue reset

If the clue is being reset, say from a new clue being read and
initialized, we should not continue to use the old location.

* Add item charges for baskets and sacks (#9212)

* timers plugin: fix teleblock timers

Co-authored-by: jordibenck <jordi.b-98@hotmail.com>

* ge plugin: add grand exchange total value

Co-authored-by: Adam <Adam@sigterm.info>

* worldmap: fix shield of arrav start points

* wintertodt plugin: add round start notification

Co-authored-by: Adam <Adam@sigterm.info>

* cache/FrameDumperTest: Don't get the wrong framemaps for a frame

* world hopper: add option to disable Hop-to menu option

* cursors plugin: add dragon dagger (p) cursor

* [maven-release-plugin] prepare release runelite-parent-1.5.28

* Resolved SoundEffectMixin merge conflict

* Merge conflict fixes & version increment