Revision 182 (#45)
* Removed duplicate files

* Refactored client for revision 182

First attempt at this so might be scuffed

* Renamed various exported rasterizer fields

* Updated injector-plugin

* RS/RL API Update

* Mixins update

* Fixed invalid static flag & added additional error debugging

* Rs-client & syntax error fixes

Project compiles again but still crashes on startup due to issues with the injected-client

* Updated to latest RL+ rs-client

Because their version of the rs-client has more refactored code and exports.

* Fixed crash on client startup

Disabling this event breaks the ToB plugin. Will fix this event if possible when I have more time.