project: upgrade to RuneLite 1.6.8 (#166)
* runelite-api: add GE search event and variables

* ItemVariations: add reverse mapping of variations

* banktags: add tag searching for ge search

* inventorygrid: re-add a delay to showing the overlay

If a player for example tries to drop an entire inventory quickly with the
overlay turned on, the inventory grid will flash for each small mouse movement
if the left mouse key is held down. This change mitigates the issue by adding a
delay option that controls the amount of time the user has to wait after
pressing it before enabling any overlay functionality. Showing the grid still
requires the mouse to be moved ~5px away from the first clicked position.

This delay just delays the activation of the overlay, as the the delay
countdown is started upon item press rather than the exit this ~5px range.
Making it respect the exit is needlessly complex as desired functionality is
reached with the current changeset.

* cluescroll plugin: Clear stale location clue hint arrows

Location clue scrolls added a hint arrow when the player was nearby one
of the clue's locations. It did not, however clear that arrow if the
locations were changed and the previously-set hint arrow location was no
longer marked. This commit updates this behavior to clear the active
hint arrow each game tick before re-adding any location hint arrows.
(and if no NPC hint arrows are set)

Fixes runelite/runelite#10370

* clue plugin: add new Slepe anagram clue

* skill calc: fix skill requirement for compost potions

* clue plugin: update solution text for deep wildy dungeon cryptic clue

* clue plugin: add pet rock and rune thrownaxe to DK clue

* Merge PR #10777 from Nightfirecat/clear-clue-scroll-on-empty-dev-command

Clear clue scroll on empty dev command

* util: Create RSTimeUnit enum

This commit adds a list of time units used in some duration
calculations, specifically client ticks and game ticks.

* plugins: Use RSTimeUnit util where applicable

This replaces uses of `Duration#minusMillis()` and `TemporalUnit`
referencing units of time defined in RSTimeUnits with their definitions.

The Timers plugin constructors are changed to accept implementations of
the `TemporalUnit` interface as they do not require any
`ChronoUnit`-specific APIs. Furthermore, the timer lengths updated use
the game tick lengths described in the OSRS wiki, which generally
closely correspond to the previous length definitions we used.

* clue plugin: update Trollweiss cryptic clue hint to include a sled

* Merge pull request #10752 from Trevor159/notification-event

client: add notification fired event

* skill calc: add long and curved bones to construction calc

* chat commands: add Bounty Hunter, Bounty Hunter Rogue, and LMS commands

* Merge pull request #10258 from BenDol/bd_itemstats_impr

itemstats: Add 'Always Show Base Stats' & 'Show Stats In Bank'

* skill calc: strip out non-digits characters

* loottracker: fix hide/show ignored items button

* item prices: show tooltips in the seed vault

* Merge pull request #10337 from sethtroll/fixagilityshortcut

Update quest start and agility shortcut coordinates for WorldMap plugin

* clue plugin: Improve Shilo Village elite coordinate clue hint

* worldmap: support multiline tooltips

* worldmap: add informative mining site tooltip

* configmanager: harden against being killed mid-save

This first writes to a temporary file, then if successful, attempts an
atomic rename

* Merge protocol and protocol-api into cache-client

Originally I was going to expand on the protocol further, but am no
longer looking at doing that.

* Thread Desktop browse and open

This should allow systems without GVFS to open the logs folder. This applies to the AppImage on non-Gnome or bare bone systems, the raw Jar, and Snap.

Co-authored-by: Adam <Adam@sigterm.info>

* slayer: fix boss task parsing to support Konar

Fixes runelite/runelite#8009

* Merge pull request #10833 from dekvall/lizards-are-reptiles-salamanders-

attackstyles: fix salamander magic attacks

* xpdrop plugin: add 'prayer' to tags

* inventory tags: add 2 additional color groups

* blast furnace: add coffer time remaining overlay

* LayoutResizableStones: fix bad script merge

* Use GE limits from item stats

* emoji: add pensive emoji

* session: add logged-in tracking

* mixins: add GE search event & fields

* gpu: merge with upstream

* Update Object IDs to 2020-02-27-rev182

* Update NPC IDs to 2020-02-27-rev182

* Update Scripts to 2020-02-27-rev182

* Bump to 1.6.8

* SanLite 1.12.5 bump

Co-authored-by: Adam <adam@sigterm.info>