feat: updates to upstream version 1.17.13 (#264)
* Merge pull request #13744 from petter/giants-den-cryptic-clue

clues: Add the new cryptic clue at Giant's Den

* discord: fix npe from menu event timing out with elapsed total time

Setting the start time to null isn't necessary, since IN_MENU is marked
shouldRestart, it always resets the start time when the status is changed to
menu. This was clearing it, forcing a presence update, and then erroring when
trying to compute the original start time. It also would have never really unset,
even if it wasn't erroring later, due to IN_MENU also being unclearable, and
so would have never been removed in the first place.

Additionally the status reset test is also wrong since it is testing the end
timestamp which has never been used by us, making the test useless.

* Merge pull request #13743 from abextm/akd

* loot tracker: check for correct menu option type when deciding event

* api: correct imcando hammer construction animation

* barbarianassault: remove dead overlay menuops

menu options are never added on dynamic widgets since they don't have
bounding boxes

* runelite-client: use menu entries for focusing world map points

The previous way had usability issues since it would cover up the close
button, consumed clicks for no reason on any icon, making it difficult
to pan around the map when zoomed out, and had no tooltips.

It previously also accessed widgets on the EDT, which is a threading
issue.

* clues: Add new A Kingdom Divided update clues

* clues: Fix clues changed with A Kingdom Divided update

* menu swapper: add enter-crypt to quick swap

This is for Sarachnis

* timers: add duration and cooldown timers for arceuus spells

Co-authored-by: Adam <Adam@sigterm.info>

* Merge pull request #13741 from Hydrox6/login-screen-new

loginscreen: add a kingdom divided login screen

* agility: Fix Shayzien basic course total xp

The final obstacle of the Shayzien basic course grants 92.2xp, not a
flat 92.

* Merge pull request #13762 from Hydrox6/skill-calc-akd

* api: add actor overhead cycle accessors

* Merge pull request #13768 from Hydrox6/venom-hitsplat

* crowdsourcing/dialog: track start & end of conversations (#13763)

* achievementdiary: Fix hard Lumbridge diary entry text (#13689)

* EmoteClue: Fix Shayzien war tent clue text

* Update Item IDs to 2021-6-23

* Update Object IDs to 2021-6-23

* Update Scripts to 2021-6-23

* Update Widget IDs to 2021-6-23

* chat icon manager: update clan rank to idx mapping

Nobody was added with key -6

* worldmap: include names for teleport icons

* feat: adds missing imports & version bump

Co-authored-by: Jordan <Nightfirecat@users.noreply.github.com>
Co-authored-by: Adam <Adam@sigterm.info>
Co-authored-by: Siraz <->
Co-authored-by: Hydrox6 <ikada@protonmail.ch>
Co-authored-by: Max Weber <mii7303@gmail.com>
Co-authored-by: Jordan Atwood <jordan.atwood423@gmail.com>
Co-authored-by: andmcadams <andmcadams@gmail.com>
Co-authored-by: SirWrain <83126316+SirWrain@users.noreply.github.com>
Co-authored-by: RuneLite Cache-Code Autoupdater <mii7303+rlccau@gmail.com>