feat: updates to upstream version 1.7.20.1 (#270)
* skill calc: add Magic Essence Potion

* loot tracker: add rogues' chest

* slayer plugin: use npchiglight for task highlights

* interact highlight: set overlay priority to high

This is to place it above npc indicators

* feat(gauntlet): hides environment overlay in boss room

* widget overlay: rebuild overlay layers when xp tracker moves

The overlay layers are sorted by position and then by priority, so
changing the position without rebuilding them is incorrect.

* widget overlay: lower priority of oxygen and hp interfaces

The client attempts to force these to always be below the xp tracker,
and having them at an equal priority can cause our overlay renderer to
fight it by laying out them above the xp tracker

* Merge pull request #14003 from Vetricci/roof-removal-prif-overrides

roof removal: add overrides for Prifddinas

* skillcalculator: Add new amethyst items (#13890)

* npc indicators: add fill color config

* woodcutting: remove respawn timers from construction trees

* ping: add macos support for icmp ping

* config manager: fsync temp config file

* text: don't remove double spaces in toJagexName

* banktags: Fix bank scroll height with separators hidden (#13723)

Prior to this change, an extra row of empty space was added to tag tabs
with items reaching the end of a row (ie. multiples of 8) which led to
scroll behavior not matching vanilla for tabs with enough items to
require scrolling. This change fixes this behavior to add extra scroll
height only when a new row is started.

In addition to the 36 pixels of height allotted to each bank item row,
an additional 4 pixels of height exists at the bottom of the bank item
container. This change ensures that a bank tag scrolled to the bottom
will have its items displaying at the exact same y values as that of a
normal bank tab.

* clientloader: fix loading classes in packages

* clientui: fix restoring sidepanel when expanding sidebar

Closing and opening the sidebar when the side panel is collapsed
currently expands the side panel to the most recently opened panel,
even if it was collapsed at the time the sidebar was closed, and also
incorrectly doesn't set the nav button as selected.

This changes the behavior of expanding the sidebar to not open the side
panel if it was not opened at the time the sidebar was closed.

* Merge pull request #14031 from Hydrox6/roofs-fixes

* Merge pull request #13907 from Hydrox6/clues-fakes

clues: remove non-existent clues

* Merge pull request #14020 from TheStonedTurtle/gaunt

* rl-client/RSProfileType: add DMM Reborn

* rl-client/ImageCapture: use RSProfileType for screenshot naming

* ItemMapping: Add combat achievement rewards variations

* slayer: Support combat achievement slayer helms

* clues: Support combat achievement slayer helms

* Merge pull request #14039 from Hydrox6/roofs-fixes

* rl-api: expose actor pose anim source setters

* CrypticClue: Fix hosidius fruit store crates highlighting (#14051)

* feat: adds missing actor overrides & version bump

* feat: updates injector & client loader

Co-authored-by: JoshS <Scriven.josh@gmail.com>
Co-authored-by: BenMortensenTTS <ttttloading@gmail.com>
Co-authored-by: Siraz <->
Co-authored-by: Adam <Adam@sigterm.info>
Co-authored-by: Jordan <Nightfirecat@users.noreply.github.com>
Co-authored-by: Vetricci <81270549+Vetricci@users.noreply.github.com>
Co-authored-by: Cyborger1 <45152844+Cyborger1@users.noreply.github.com>
Co-authored-by: Max Weber <mii7303@gmail.com>
Co-authored-by: Jordan Atwood <jordan.atwood423@gmail.com>
Co-authored-by: TheStonedTurtle <29030969+TheStonedTurtle@users.noreply.github.com>