feat: updates to upstream version 1.7.24 (#272)
* overlay: only layout widget overlays if the preferred location is set

With the new steam client updates, the client tries to aggressively
move around several of the widget overlay components, including xp
tracker, and several hp bars. This tends to fight with the layouting we
already do with the widgets.

This changes the widgets to only be layouted if the overlay has a
preferred location or position set. Otherwise, it will offset the snap
corner bounds by where the client has positione the widget.

It is still possible to get the client ui to fight the widget overlay by
moving the overlays around manually, but it should no longer happen in
the default positioning of the widgets.

* world hopper: bubble up mouse events from activity label

This fixes hopping and mouseover on the activity label not working

* timetracking: don't load notification icons for every panel

* Merge pull request #13908 from Hydrox6/sign-in-wording

client: update various wordings to use "signed in/out" for RL accounts

* Merge pull request #13702 from SkylerPIlot/Skill_Calc_Addition

* Merge pull request #14074 from Hydrox6/loot-tracker-ea

* Merge pull request #14058 from Nightfirecat/fix-ardy-monastery-roofs

roof removal: Fix Ardougne monastery

* Merge pull request #14042 from Hydrox6/roof-removal-instances

roof removal: properly account for instances when applying overrides

* loot tracker: add world id to loot record for dmmt

* Merge pull request #14073 from Hydrox6/xp-globes-correct-bounds

xp globes: account for progress arc width when deciding overlay bounds

* mining: Add barronite ore and deposits

* skillcalculator: Add Camdozaal activities

* worldmap: Add mining spots in Ruins of Camdozaal

* ItemMapping: Add Tome of Water and Master Scroll Book

* logback: add duplicate exception filter

This only applies to marked logged messages for overlay rendering and event subscribers, which are common sources of exceptions in 3rd party plugins

* wiki: move icon to match vanilla position

* ui: force heavyweight tooltips on macos

Lightweight components do not render correctly over AWT canvases on
MacOS. The popup factory API to override this is not available on 8, so
add an 11-specific popup factory to override it by using a multi-release
jar.

* npc highlight: fix npe rendering minimap names on npcs with no name

* Move npc highlight overlay to core

This allows the other plugins which depend on the npc overaly to work
when npc highlight plugin is off, while still letting them use the
consistent highlight style of the central overlay.

* api: make getCanvasTilePoly work on multi-tile npcs

* bank tags: Fix tag tab tab scroll height (#13761)

* runecraft: use npc overlay service

* corp: use npc overlay service

* external plugins: require assertions when loading builtin plugins

* itemstats: Add jangerberries prayer point restore (#14101)

* feat(mixins): updates for api changes

* fix: compilation errors

* Update Scripts to 2021-9-9

* Update Item IDs to 2021-9-9

* Update Item variations to 2021-9-9

* Update Object IDs to 2021-9-9

* Update NPC IDs to 2021-9-9

* npc minimap overlay: use unmodified highlight color

The darker() is originally from the npc indicators minimap overlay, but other plugins now using this such as impling plugin do not expect this.

If we require the minimap colors to be different later we should add a separate property specifically for it.

* chore: version bump

* feat: various mixin updates

* npc overlay: set lower priority on npc change and despawn handlers

Similar to spawn, these need to run after plugins for most of the highlight functions to work correctly

* widget overlay: correct raids points overlay position

* music: check game state before applying sliders on startup

This accesses vars, which typically cannot be accessed when not logged in, or in early startup

* npc overlay: clear highlighted npcs on logout

Despawn events don't fire for npcs on logout, copied from the npc highlight plugin

* client: update jna to 5.9.0

This is required for macos aarch64 support

* feat: updates bootstrap

* rl-client: add plugin conflict support

* feat: more mixin fixes

* api: add model vertex normals

* api: add scene tile shapes, underlay, and overlay ids

* feat: more api and mixin changes

Co-authored-by: Adam <Adam@sigterm.info>
Co-authored-by: Max Weber <mii7303@gmail.com>
Co-authored-by: Jordan <Nightfirecat@users.noreply.github.com>
Co-authored-by: Adam Davies <acdavies001@gmail.com>
Co-authored-by: pilino1234 <pilino@posteo.de>
Co-authored-by: Ryan <ryangreen105+siemens@gmail.com>
Co-authored-by: Siraz <->
Co-authored-by: geheur <41499327+geheur@users.noreply.github.com>
Co-authored-by: Vic Segers <32434088+VicSegers@users.noreply.github.com>
Co-authored-by: RuneLite Cache-Code Autoupdater <mii7303+rlccau@gmail.com>