feat: updates to upstream version 1.8.3 (#282)
* ground items: Fix despawn timer in clan hall (#14358)

* banktags: remove extra space in chatbox prompt

* banktags: improve bank scrollbar updates

With tag tab tab or "Remove tab separators", the scrollbar height needs to be
adjusted to account for either many tag tabs (which requires scrolling to see)
or the hiding of the tab separators. This is because we artifically filter the
bank results without keeping the bank in searching mode, causing the vanilla
code which usually computes the scroll height to be incorrect.

The existing code does this by manually invoking [proc,update_scrollbar] later
with the recomputed scroll height. Now that invokeLater runs next tick and not
next frame, this leaves several frames with the scrollbar the incorrect size.

The desired scroll height is passed to [proc,bankmain_finishbuilding] in
argument int12, so instead we recompute the scroll height and overwrite the
script argument prior to execution.

* banktags: remove setBankScroll event

This hasn't worked in awhile since it was assining the scroll value to the
wrong variable. Now we are computing the scroll in onScriptPreFixed for
[proc,bankmain_finishbuilding], so it is no longer necessary here.

* clues: Accept Tome of water when requiring water runes (#14361)

* Merge pull request #11321 from Hydrox6/clues-runepouch

* clues: remove unnecessary item array copy

The array returned by getItems() is always a new array and so does not require copying

* clues test: remove unnecessary stubbing

* chat filter: Ignore character accents for matching

This lets plain latin-character filters to match messages with accents
and diacritics which are not easily typed on all keyboard layouts.

Co-authored-by: Jordan Atwood <jordan.atwood423@gmail.com>

* Merge pull request #14369 from LlemonDuck/spec-counter-ignore-vetion-hellhounds

* gpu: add sync mode config

This allows the additional OFF mode which has the client manage its own
frame rate target without the use of vsync. On lower end systems which
have a hard time delivering a frame in the vsync interval this can give
a noticible performance improvement since it doesn't stall the cpu
waiting for the next vblank when a frame is skipped

* avoid boxing `Comparator.comparing`s

* timestamp: use steam client chat timestamp support

The chat building scripts have variables now for message timestamp, so we can just assign our timestamp to those instead.

* timestamp: immediately apply color config changes

* Update Scripts to 2021-11-17

* Update Item IDs to 2021-11-17

* Update Item variations to 2021-11-17

* menu swapper: add teleports swap to teleport swap

This is for the max cape teleports option

* screenshot: update duel arena messages

* chat commands: update duel arena messages

* achievement diary: remove the queen of thieves quest requirement

* feat: updates mixins & bumps version

Co-authored-by: Alex <Alex@romanink.net>
Co-authored-by: Adam <Adam@sigterm.info>
Co-authored-by: simeonlg <simeonlg@users.noreply.github.com>
Co-authored-by: Jordan <Nightfirecat@users.noreply.github.com>
Co-authored-by: Jordan Atwood <jordan.atwood423@gmail.com>
Co-authored-by: Max Weber <mii7303@gmail.com>
Co-authored-by: RuneLite Cache-Code Autoupdater <mii7303+rlccau@gmail.com>
Co-authored-by: Siraz <->
Co-authored-by: Brad Rammel <bmrammel@gmail.com>