ge plugin: compact buy offer examine text
The examine text, buy limit, trade price, and fee info does not fit on
the interface anymore since the addition of the fee info. This compacts
the examine text as well as places the limit / trade price on the same
line in order to prevent pushing the fee off of the bottom of the
border.