Fix race with client startup and low detail/chat icon loading
Since c1aa0c4898900361593e423792b475c08f141e49 moved client startup to
be much earlier, the plugins can now start after the client is already
at the login screen. This causes the low detail plugin and chat icon
manager to not initialize correctly as they are waiting on the event
for the login screen.