feat: updates to upstream 1.8.12 (#292)
* fairy ring: add Yu'biusk

* widget overlay: Don't draw empty wilderness K/D box

The Wilderness and PVP kill-death information box is created while in
these areas regardless of whether the setting (configured via the notice
board at the Edgeville bank) is enabled to show the text, meaning the
widget contains only empty text widgets when the setting is disabled
rather than being null, causing a bounding box to still be drawn and
affecting other snapped widget layout. This commit adds a child class of
WidgetOverlay specific to this widget and prevents it from being drawn
when the setting to show this information is disabled.

* gpu: apply hsl override to flat shade faces

The behavior of this was changed in 203 to include flat shade faces

* Update dnschange url

The old url now tries to sell users a vpn

* notifier: set app name when using notify-send

If a notification is sent while the screen is locked, it says
"notify-send" instead of "RuneLite".

* linkbrowser: prefer xdg-open over Desktop.open and browse

Desktop.open/Desktop.browse doesn't work on Linux with gtk3

See JDK-8275494

* loot tracker: add hallowed sack

* cache: refactor ModelLoader.decodeOldFormat variable names

* clues: add Jimmy Dazzler clue

* clues: add Yu'biusk clue

* screen markers: fix visibilityLabel tooltip

The tooltip was not correctly toggling between show/hide when the visbility was toggled

* screen markers: add a tooltip to border thickness spinner

* screen markers: remove fill toggle

This predates the alpha colorpicker, so it sort of made sense at the time, but currently the same effect can be achieved by just changing the fill color alpha. Additionally the toggle doesn't really toggle it but instead changes the alpha between 0 and 75 - and not the alpha that is chosen by the picker.

* screen markers: make ScreenMarkerRenderable implement RenderableEntity

This component is never added to the overlay componenet renderer so implementing LayoutableRenderableEntity was just unnecessary

* screen markers: add marker labels

This adds a toggle to display the marker name as a label above the
marker

Co-authored-by: Skretzo <53493631+Skretzo@users.noreply.github.com>

* jshell: fix run/clear tooltips

* feat: updates archive-patcher to 1.2

* Update Item IDs to 2022-2-23

* Update Item variations to 2022-2-23

* Update Object IDs to 2022-2-23

* Update NPC IDs to 2022-2-23

* Update Scripts to 2022-2-23

* chore: bumps version

* chore: bumps sanlite version

Co-authored-by: testing-ongithub <96100526+testing-ongithub@users.noreply.github.com>
Co-authored-by: Jordan Atwood <jordan.atwood423@gmail.com>
Co-authored-by: Adam <Adam@sigterm.info>
Co-authored-by: Jeremy Plsek <jeremyplsek@gmail.com>
Co-authored-by: Christos-Apostolidis <72702348+Christos-Apostolidis@users.noreply.github.com>
Co-authored-by: Nakst <17091026+nakst@users.noreply.github.com>
Co-authored-by: Skretzo <53493631+Skretzo@users.noreply.github.com>
Co-authored-by: Siraz <->
Co-authored-by: RuneLite Cache-Code Autoupdater <mii7303+rlccau@gmail.com>