feat: updates to runelite version 1.8.13 (#293)
* gpu: Fix loading gluegen natives on some Windows configurations

This commit works around a bug with loading natives when the user's home
folder contains special characters such as `&`, `^` or `!`.

When Gluegen loads native libraries on Windows, it uses a temporary
directory located in `%appdata%` under the user's home directory, and
Gluegen checks whether files can be executed from this temporary
directory. To perform this check, it writes an executable to the
directory, executes it, and ensures that the return code is zero. The
executable file it writes is by default a `.bat` file, but it can be
told to write an `.exe` file instead by supplying this VM argument:
`-Djogamp.gluegen.UseNativeExeFile=true`.

The default behaviour of writing and executing a `.bat` file is broken
when the path to the temp folder contains a special character, due to a
convoluted trail of calls that leads to undefined behaviour:
- Gluegen attempts to execute the `.bat` file by calling `Runtime.exec()`:
  `Runtime.exec(new String[] { "...absolute path...bat" }, null, null)`
- OpenJDK 11.0.8 passes the command to `ProcessBuilder`, which calls
  `ProcessImpl.start(...)`, which creates a new `ProcessImpl` instance.
- In the `ProcessImpl` constructor, there are two different character
  escape modes:
  - **Legacy mode**, which allows ambiguous commands.
  - **Non-legacy mode**, which does more rigorous escaping.
- Legacy is the default mode as long as `System.getSecurityManager()`
  returns `null`, which is the case for RuneLite currently.
  This mode can be changed by supplying the following VM argument:
  `-Djdk.lang.Process.allowAmbiguousCommands=false`.
  - Incidentally this also fixes the issue of loading natives with
    special characters in the path, however it applies to *all*
    `Runtime.exec()` calls.
- In legacy mode, the command is wrapped in quotes if it contains either
  a space or tab character. The command is then passed along to the
  native function `Java_java_lang_ProcessImpl_create`, which eventually
  passes it along untouched to Microsoft's `CreateProcessW` function as
  the `lpCommandLine` argument.
- In the documentation for this function, it is mentioned that an
  interpreter (i.e. `cmd.exe`) is required in order to run a `.bat`
  file. It is not specified what will happen if you instead pass the
  path to a `.bat` file directly to the function, which is what Gluegen
  ends up doing. However, `CreateProcessW` *does* support supplying the
  path to an `.exe`, which Gluegen will do when
  `jogamp.gluegen.UseNativeExeFile` is set to `true`.

It is unclear why supplying a `.bat` file normally works, yet breaks
when special characters are in the path. It seems like it should not
work in the first place.

This might help with #6509 and #14180.

Some other special characters, like Cyrillic letters, are more
fundamentally broken in RuneLite. This patch only slightly helps with
that.

* clues: fix Guardian mummy capitalization

* slayer: Fix name matching

The Slayer plugin highlights target monsters based on their name rather
than NPC ID, as many common monsters (skeletons, zombies, etc.) have
nearly endless variations for different models and combat levels.
Previously, this name matching was done via a simple
`String#contains()`, which led to some incorrect matches such as pirates
being highlighted while on rat tasks and Jonny the beard being
highlighted while on bear tasks.

This commit changes matching to use regex to match string boundaries or
whitespace at either end of the task string, ensuring these substring
matches can only happen when word breaks occur. The only known existing
case where this would apply is for baby dragons and brutal dragons,
which are valid alternatives for their respective chromatic dragon
tasks.

* clientui: run PluginPanel#onDeactivate when switching panels

* cache: script: use linkedhashmap for switch map

The switch maps are iterated in the assembler and disassembler and the generated code depends on the iteration order

* cache: script: remove unused import

* item manager: fix active price threshold calculation

This was meant to keep items within 5x of their Jagex price, but the
logic was only correct if activePriceThreshold <= 1.0.

* idle notifier: disable by default

* cache: rename texture animation direction and speed

* gpu: move texture animation to gpu

* gpu: fix anim array indexes with sparse texture array

Texture 54 doesn't exist and was causing everything after it to have its index off by 1

* loottracker: split panel construction into methods

This moves the overall and actions panel creation into separate methods

* loottracker: replace getTotalPrice with stream

* loottracker: hoist box emptyborder from loop

* config manager: post RuneScapeProfileChanged when new profiles are created

* loot tracker: store loot in config

Since loot is now aggregated, the data is little enough to store in
config. This allows loot to persist between sessions even when not
logged in.

* runelite: disable option parser abbreviations

* api: add methods to get selected item

* config manager: run shutdown hook after plugins

* gpu: don't animate textures when loading

* api: add isFollower to NPCComposition

* item identification: add goblin and magic ess potions

* item identification: add rejuvenation potion

* Add barbarian rod anim to fishing and idle notifier

* api: expand item api for inventory model replacement

* client: add plugin sideloading

This supports sideloading plugins out of the sideloaded-plugins
directory when run in developer mode

* feat: updates rs-api

Co-authored-by: Hooder <ahooder@protonmail.com>
Co-authored-by: Cyborger1 <45152844+Cyborger1@users.noreply.github.com>
Co-authored-by: Jordan Atwood <jordan.atwood423@gmail.com>
Co-authored-by: Tony Wang <74754989+hwang-pku@users.noreply.github.com>
Co-authored-by: Adam <Adam@sigterm.info>
Co-authored-by: David Luong <davidluong98@gmail.com>
Co-authored-by: Faycal <b.fayccal@gmail.com>
Co-authored-by: Siraz <->