feat: updates to upstream version 1.8.20 (#301)
* timers: Add Shadow Veil protection detection (#14687)

* item stats: Add some missing beverages (#14412)

* skill calculator: Fix herblore typos (#14324)

* rl-api, groundmarkers: handle instance plane conversion correctly

* overlay: better handle picking up overlapping overlays

findManagedOverlay needs to iterate in roughly the reverse drawing
order. Also adjust the moveable tli layers to be by priority in render
oder.

* menu swapper: removetags object/npc names

* menu swapper: removetags npc name on reset swap

* clientui: setResizable after setVisible

Prevents an issue with clients being resizable with toolbar-less means in some environments (e.g. double clicking title bar on macos).

Honestly I'm not entirely sure why this is ignored if the window isn't visible, rather than being propagated to the window manager when the window is shown, but I guess Swing isn't that smart (like always).

I originally found the information on [this SO answer](https://stackoverflow.com/a/14884056) and it seemed to resolve the issue.

Closes #14235

* itemcharges: add imbued ring of wealth

* widget overlay: change minimap position back to top_right

Otherwise layouted overlays don't get layouted around the minimap

* overlay: add movable and snappable overlay flags

This merges DETACHED and DYNAMIC into one position, DYNAMIC. And, allows
detached/dynamic overlays to be snappable.

* overlay: disallow transformPosition with dynamic/tooltip

* overlay: set g2d clip per layer

* overlay: worldmap: prefer rect2d as g2d clip shape

* overlay: do not move snap points backwards

because WidgetOverlays in their default position render in the vanilla
position it can cause other overlays to occlude previous overlays when
they should not

* chat commands: Update clear word and clear line defaults

Ctrl + backspace is an overwhelmingly common keybind for clearing the
word before the cursor. Meanwhile, clearing the current line is not as
common a keybind, so having an unset default is more sensible here for
users to decide for themselves what keybind they would prefer.

* devtools: allow pgup/down to cycle through previous commands

Co-authored-by: Max Weber <mii7303@gmail.com>

* overlay: use hovered overlay for picking managed overlay

Hovered overlay already accounts for render order, with the exception of
the vanilla WidgetOverlays which are drawn prior to any of our overlays
being drawn. Instead we order them in render-order, which works for the
dynamic-layer ones.

The minimap must be top_right so that our overlays layout around it
correctly, which currently causes it to be picked up first over dynamic
overlays due to the way overlay sorting is done.

* Update Item IDs to 2022-5-18

* Update Scripts to 2022-5-18

* Update Item variations to 2022-5-18

* Update Object IDs to 2022-5-18

* Update NPC IDs to 2022-5-18

* fairyring: add isle of souls ring

* feat: updates injector

Co-authored-by: ProjectileRage <19807407+ProjectileRage@users.noreply.github.com>
Co-authored-by: Jonathan Forscher <jonforscher@gmail.com>
Co-authored-by: Matthew C <Chapman.L.Matthew@gmail.com>
Co-authored-by: Max Weber <mii7303@gmail.com>
Co-authored-by: Adam <Adam@sigterm.info>
Co-authored-by: LlemonDuck <napkinorton@gmail.com>
Co-authored-by: Stefan Zopfi <zopfistefan@gmail.com>
Co-authored-by: Jordan Atwood <nightfirecat@nightfirec.at>
Co-authored-by: RuneLite Cache-Code Autoupdater <mii7303+rlccau@gmail.com>
Co-authored-by: Siraz <->