feat: updates to upstream version 1.8.21 (#302)
* overlay manager: remove spurious position reset logs

* gpu: initialize scene uploader scene id with nanotime

Multiple gpu plugins initializing the scene id to the current time can cause conflicts. This makes that less likely.

* api: add setForcedPosition widget method

* overlay: prevent moving non-movable overlays

* prayer: don't return dimension for dynamic overlays

The returned dimension is only used for layoutable or movable overlays

* api: pass menu entry to menu add event and forward accessors

* events: add tostring/equalsandhashcode to menuoption clicked

* api: add npc accessor to menuentry

* api: add player accessor to menuentry

* item stats: Add missing consumables

* attackstyles: add weapon types 28 and 29

* fix race loading modicons

It can't be assumed the array passed to setModIcons is still correct in
the async image callbacks. Instead, fetch the current modicons array in
the callback and use that.

* api: add player id accessor

* Update NPC IDs to 2022-5-26

* Update Scripts to 2022-5-26

* item stats: Clean up saradomin brew code

* item stats: add support for stat boosts with a cap

* item stats: add ancient brew

* item stats: Use variables for reused potion effects

* item stats: Add mixed potions

* item stats: Remove combo primaries

This field never had any real use from the beginning, and has only
caused confusion for contributors adding item stat change definitions.

* item stats: Fix jangerberries stats

* timers: Track home and minigame teleports using vars (#14842)

* status bars: Override hitpoints and prayer max values in LMS

While playing Last Man Standing, your hitpoints and prayer values behave
as though their maximum values were 99 instead of their actual max
values.

* fishing: Add frog spawn spot NPC (#14659)

* party: remove Discord requirement

* party: use passphrases for party ids

These are easier to type and share than uuids.

* spec counter: add spec drops

Co-authored-by: Adam <Adam@sigterm.info>

* party: remove overlay

This is rarely used because the information is duplicated in the panel, and not very useful. Also the toggling behavior is confusing to people.

* spec counter: add devmode check for spec command

* special counter: Fix test

* hunter: Remove unused lastActionTime field

This field was originally used to show a catch rate overlay, but all its
uses were removed in 76fbab60fc742c95c526ce0cb8a371f8d5d65369 when the
overlay was removed due to unreliable behavior.

* gpu: fix debug mode on macos

debug mode aggressively checks and throws whenever a gl error occurs.

The glBlitFramebuffer call is bizzare because with the alpha channel
enabled on the renderbuffer it appears to work correctly despite
causing an INVALID_OPERATION error. We have no use for the alpha channel
since the dst alpha value is always 1, so it can just be disabled on macos.

* worldmap: add akp and bjp fairy rings

* item identification: Add Desert Treasure diamonds (#14764)

* ClientLoader: don't fail patching to hidden files

FileOutputStream (by design) passes FILE_ATTRIBUTE_NORMAL and
CREATE_ALWAYS to CreateFile, which causes it to fail with access denied
if it is opening a FILE_ATTRIBUTE_HIDDEN or FILE_ATTRIBUTE_SYSTEM file

See JDK-8047342

* SessionManager: don't fail to login with a existing hidden session file

* feat: updates rs-api & mixins

* fix: corrects bootstrap artifact path issue & bumps version

Co-authored-by: Adam <Adam@sigterm.info>
Co-authored-by: emerald000 <emerald000@hotmail.com>
Co-authored-by: RuneLite Cache-Code Autoupdater <mii7303+rlccau@gmail.com>
Co-authored-by: Jordan Atwood <nightfirecat@nightfirec.at>
Co-authored-by: Hydrox6 <ikada@protonmail.ch>
Co-authored-by: superiorser9 <58041466+superiorser9@users.noreply.github.com>
Co-authored-by: vmarlowe <77749581+vmarlowe@users.noreply.github.com>
Co-authored-by: Siraz <->
Co-authored-by: Hexagon <hexagon@fking.work>
Co-authored-by: David Luong <davidluong98@gmail.com>
Co-authored-by: Max Weber <mii7303@gmail.com>